refactor: like 관련 기능을 curation 으로 이동
